requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
aspose-words==23.3.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.2.0
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-admin-honeypot==1.1.0
django-cleanup==6.0.0
django-crispy-forms==1.13.0
django-hashid-field==3.3.7
django-js-asset==2.1.0
django-session-timeout==0.1.0
django-user-visit==0.4
django-xhtml2pdf==0.0.4
fonttools==4.47.0
frozenlist==1.4.1
future==0.18.2
gunicorn==20.1.0
hashids==1.3.1
html5lib==1.1
idna==3.6
importlib-resources==6.1.1
jsonpickle==3.0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.3.0
matplotlib==3.8.2
multidict==6.0.5
numpy==1.26.2
oscrypto==1.3.0
packaging==23.2
pango==0.0.1
pdfkit==1.0.0
pillow==11.0.0
psycopg2==2.8.6
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==5.0.1
PyPDF2==1.26.0
pyphen==0.16.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2021.3
PyYAML==6.0.2
qrcode==8.0
qtoml==0.2.4
referencing==0.33.0
regex==2021.11.10
reportlab==4.2.5
requests==2.31.0
rpds-py==0.17.1
sanskrit-data==0.8.13
sanskrit-text==0.2.3
simple-colors==0.1.5
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.3.0
toml==0.10.2
tqdm==4.62.3
twilio==9.0.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ua-parser==0.16.1
uritools==4.0.3
urllib3==2.2.1
user-agents==2.2.0
validators==0.20.0
weasyprint==62.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.16
yarl==1.9.4
YURL==1.0.0
zipp==3.17.0
zopfli==0.2.3.post1
